requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-firestore
google-cloud-bigquery==1.10
google-cloud-core==0.28.1
pandas
